+++ themes/themeOrange/views/user/views/auth/login_modal.php
@@ -2,9 +2,11 @@
 
 /* Added Language Switcher */
 
+use humhub\libs\Html;
 use humhub\modules\user\models\forms\Login;
 use humhub\modules\user\models\Invite;
 use yii\captcha\Captcha;
+use yii\helpers\ArrayHelper;
 use yii\widgets\ActiveForm;
 use yii\helpers\Url;
 use humhub\modules\user\widgets\AuthChoice;
@@ -13,6 +15,7 @@ use humhub\modules\user\widgets\AuthChoice;
 /* @var $model Login */
 /* @var $invite Invite */
 /* @var $info string */
+/* @var $passwordRecoveryRoute string|array|null */
 ?>
 <div class="modal-dialog modal-dialog-small animated fadeIn">
     <div class="modal-content">
@@ -71,13 +74,26 @@ use humhub\modules\user\widgets\AuthChoice;
                             </button>
 
                         </div>
+                        <?php if ($passwordRecoveryRoute) : ?>
                         <div class="col-md-8 text-right">
                             <small>
-                                <a id="recoverPasswordBtn" href="#" data-action-click="ui.modal.load" data-action-url="<?= Url::to(['/user/password-recovery']) ?>">
-                                    <?= Yii::t('UserModule.auth', 'Forgot your password?') ?>
-                                </a>
+                                <?= Html::a(
+                                    Html::tag('br') . Yii::t('UserModule.auth', 'Forgot your password?'),
+                                    $passwordRecoveryRoute,
+                                    ArrayHelper::merge([
+                                        'id' => 'recoverPasswordBtn',
+                                    ], is_array($passwordRecoveryRoute) ? [
+                                        'data' => [
+                                            'action-click' => 'ui.modal.load',
+                                            'action-url' => Url::to($passwordRecoveryRoute),
+                                        ]
+                                    ] : [
+                                        'target' => '_blank',
+                                    ]),
+                                ) ?>
                             </small>
                         </div>
+                        <?php endif; ?>
                     </div>
 
                     <?php ActiveForm::end(); ?>